+++ packages/pbr/src/pbr.fs.glsl
@@ -7,6 +7,8 @@
 #include <normal_share>
 #include <color_share>
 #include <worldpos_share>
+#include <refract_share>
+
 
 #ifdef ALPHA_MASK
 
@@ -102,12 +104,6 @@ uniform float u_clearCoatRoughness;
 #endif
 
 uniform vec2 u_resolution;
-
-#ifdef HAS_PERTURBATIONMAP
-    uniform sampler2D u_perturbationSampler;
-    uniform float u_perturbationUOffset;
-    uniform float u_perturbationVOffset;
-#endif
 
 
 // structures
@@ -591,11 +587,11 @@ float getLuminance(vec3 color)
 }
 
 void main() {
-
+    vec2 uv = vec2(0., 0.);
     #ifdef  USE_SCREENUV
-          vec2 uv = getScreenUv();
-    #else
-          vec2 uv = v_uv;
+        uv = getScreenUv();
+    #elif defined( O3_HAS_UV ) || defined( O3_NEED_UV ) || defined( O3_HAS_ENVMAP ) || defined( O3_HAS_LIGHTMAP )
+        uv = v_uv;
     #endif
 
     vec3 normal = getNormal(uv);
@@ -834,18 +830,10 @@ void main() {
 
     #ifdef GAMMA
         float gamma = 2.2;
-        gl_FragColor.rgb = pow(gl_FragColor.rgb, vec3(1.0/gamma));
+        gl_FragColor.rgb = pow(gl_FragColor.rgb, vec3(1.0 / gamma));
     #endif
 
-    #ifdef HAS_PERTURBATIONMAP
-        vec4 screenColor = texture2D(u_perturbationSampler, getScreenUv() + vec4(u_viewMat * vec4(normal, 1.)).xy * vec2(u_perturbationUOffset, u_perturbationVOffset));
-        if (gl_FragColor.a < 1.) {
-           // mock alpha blend
-           gl_FragColor *=  gl_FragColor.a;
-           gl_FragColor += (1.0 - gl_FragColor.a)  * screenColor;
-        }
-    #endif
-
+    #include <refract_frag>
     #include <fog_frag>
 
 }