+++ playground/src/custom-material/index.js
@@ -1,13 +1,13 @@
-import {vec3, vec4} from '@alipay/o3-math';
-import {Logger, ClearMode} from '@alipay/o3-base';
-import {Engine} from '@alipay/o3-core';
-import {AGeometryRenderer} from '@alipay/o3-geometry';
-import {ADefaultCamera} from '@alipay/o3-default-camera';
-import {SphereGeometry, CuboidGeometry} from '@alipay/o3-geometry-shape';
-import {ResourceLoader, Resource} from '@alipay/o3-loader';
-import {AAmbientLight, ADirectLight} from '@alipay/o3-lighting';
-import {HatchingMaterial} from './HatchingMaterial';
-import {AOrbitControls} from '@alipay/o3-orbit-controls';
+import { vec3, vec4 } from '@alipay/o3-math';
+import { Logger, ClearMode } from '@alipay/o3-base';
+import { Engine } from '@alipay/o3-core';
+import { AGeometryRenderer } from '@alipay/o3-geometry';
+import { ADefaultCamera } from '@alipay/o3-default-camera';
+import { SphereGeometry, CuboidGeometry } from '@alipay/o3-geometry-shape';
+import { ResourceLoader, Resource } from '@alipay/o3-loader';
+import { AAmbientLight, ADirectLight } from '@alipay/o3-lighting';
+import { HatchingMaterial } from './HatchingMaterial';
+import { AOrbitControls } from '@alipay/o3-orbit-controls';
 
 
 Logger.enable();
@@ -18,20 +18,20 @@ let rootNode = scene.root;
 //-- create camera
 let cameraNode = rootNode.createChild('camera_node');
 let camera = cameraNode.createAbility(ADefaultCamera, {
-  canvas: 'o3-demo', position: [0, 0, -10], target: [0, 0, 0]
+  canvas: 'o3-demo', position: [0, 0, -10], target: [0, 0, 0],
 });
 camera.setClearMode(ClearMode.SOLID_COLOR, [1, 1, 1, 1]);
 
-let controler = cameraNode.createAbility(AOrbitControls, {canvas: document.getElementById('o3-demo')});
+let controler = cameraNode.createAbility(AOrbitControls, { canvas: document.getElementById('o3-demo') });
 controler.autoRotate = false;
 controler.minDistance = 4;
 controler.maxDistance = 50;
 
 //-- create lights
-let light2 = rootNode.createChild("light2");
+let light2 = rootNode.createChild('light2');
 let dirLgt = light2.createAbility(ADirectLight, {
   color: vec3.fromValues(1, 1, 1),
-  intensity: 0.8
+  intensity: 0.8,
 });
 light2.position = [-10, 20, -15];
 light2.lookAt([0, 0, 0], [0, 1, 0]);
@@ -39,14 +39,32 @@ light2.lookAt([0, 0, 0], [0, 1, 0]);
 let resourceLoader = new ResourceLoader(engine);
 
 let textureRes = [];
-for (let i = 0; i < 6; i++) {
-  textureRes.push(
-    new Resource('hatch' + i, {
-      type: 'texture',
-      url: require(`./hatch_${i}.jpg`)
-    })
-  );
-}
+textureRes.push(
+  new Resource('hatch0', {
+    type: 'texture',
+    url: '/static/texture/custom-material/hatch_0.jpg',
+  }),
+  new Resource('hatch1', {
+    type: 'texture',
+    url: '/static/texture/custom-material/hatch_1.jpg',
+  }),
+  new Resource('hatch2', {
+    type: 'texture',
+    url: '/static/texture/custom-material/hatch_2.jpg',
+  }),
+  new Resource('hatch3', {
+    type: 'texture',
+    url: '/static/texture/custom-material/hatch_3.jpg',
+  }),
+  new Resource('hatch4', {
+    type: 'texture',
+    url: '/static/texture/custom-material/hatch_4.jpg',
+  }),
+  new Resource('hatch5', {
+    type: 'texture',
+    url: '/static/texture/custom-material/hatch_5.jpg',
+  }),
+);
 resourceLoader.batchLoad(textureRes, (err, res) => {
 
   //-- create sphere