+++ playground/src/effect-trail-renderer/index.js
@@ -17,9 +17,9 @@ const rootNode = scene.root;
 // 在场景中创建相机节点、配置位置和目标方向
 const cameraNode = rootNode.createChild('camera_node');
 let camera = cameraNode.createAbility(ADefaultCamera, {
-  canvas: 'o3-demo', position: [0, 0, 10], target: [0, 0, 0]
+  canvas: 'o3-demo', position: [0, 0, 10], target: [0, 0, 0],
 });
-let controler = cameraNode.createAbility(AOrbitControls, { canvas: document.getElementById('o3-demo')});
+let controler = cameraNode.createAbility(AOrbitControls, { canvas: document.getElementById('o3-demo') });
 
 // 控制 node 延蝴蝶曲线运动
 class AButterFlyMove extends NodeAbility {
@@ -33,7 +33,7 @@ class AButterFlyMove extends NodeAbility {
   update(deltaTime) {
     let time = (this.engine.time.timeSinceStartup + this.startTime) * this.trend;
     let t = time * 0.001;
-    let d = (Math.exp(Math.cos(t)) - 2*Math.cos(4*t) + Math.pow(Math.sin(t/12), 5)) * 0.8;
+    let d = (Math.exp(Math.cos(t)) - 2 * Math.cos(4 * t) + Math.pow(Math.sin(t / 12), 5)) * 0.8;
     let x = Math.sin(t) * d;
     let y = Math.cos(t) * d;
     let z = -1;
@@ -48,23 +48,23 @@ class AButterFlyMove extends NodeAbility {
 
 const techRes = new Resource('image', {
   type: 'texture',
-  url: require('./tail.png')
+  url: '/static/texture/effect-trail-renderer/tail.png',
 });
 
 const props = {
   material: new TrailMaterial('trail_mtl'),
-  stroke: 0.2
+  stroke: 0.2,
 };
 
 const resourceLoader = new ResourceLoader(engine);
-resourceLoader.load(techRes, (err,res) => {
+resourceLoader.load(techRes, (err, res) => {
   const texture = res.asset;
   props.material.setValue('u_texture', texture);
 });
 
 // 在场景中创建 butterfly 节点
 function createButterflyTail(name, startTime) {
-  const trail = rootNode.createChild("name");
+  const trail = rootNode.createChild('name');
   const trailRenderer = trail.createAbility(ATrailRenderer, props);
   const move = trail.createAbility(AButterFlyMove, { startTime });
 }